% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_area_mst.R
\name{Clust_area_mst}
\alias{Clust_area_mst}
\title{clustered area level model, mst algorithm based on admm}
\usage{
Clust_area_mst(
  dom,
  y,
  x,
  index_d,
  vardir,
  graph0,
  gamma0,
  rho = 1,
  gam = 3,
  lam = 0.5,
  lam0 = 0.001,
  maxiter = 500,
  tol = 1e-04
)
}
\description{
clustered area level model, mst algorithm based on admm
}
