% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_area_pw2.R
\name{Clust_area_pw2}
\alias{Clust_area_pw2}
\title{clustered area level model pairwise}
\usage{
Clust_area_pw2(
  indexy,
  y,
  x,
  index_d,
  vardir,
  nu = 1,
  gam = 3,
  lam = 0.5,
  lam0 = 0.001,
  maxiter = 500,
  tol = 1e-04,
  seed = 1234,
  method = "kmeans",
  k0 = 10
)
}
\description{
clustered area level model pairwise
}
