% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_clust_area2.R
\name{Est_clust_area2}
\alias{Est_clust_area2}
\title{wrap up function pairwise
@import sae}
\usage{
Est_clust_area2(
  dom,
  y,
  x,
  index_d,
  vardir,
  lamvec,
  nu = 1,
  gam = 3,
  lam0 = 0.001,
  maxiter = 500,
  tol = 1e-04,
  seed = 1234
)
}
\description{
wrap up function pairwise
@import sae
}
