% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_clust_area_mst.R
\name{Est_clust_area_mst}
\alias{Est_clust_area_mst}
\title{based on given data, give the area level model estimator}
\usage{
Est_clust_area_mst(
  dom,
  y,
  x,
  index_d,
  vardir,
  graph0,
  gamma0,
  lamvec,
  rho = 1,
  gam = 3,
  lam0 = 0.001,
  maxiter = 200,
  max_inner = 10,
  tol = 1e-04
)
}
\description{
based on given data, give the area level model estimator
}
