% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_clust_area_mst2.R
\name{Est_clust_area_mst2}
\alias{Est_clust_area_mst2}
\title{based on given data, give the area level model estimator new algorithm}
\usage{
Est_clust_area_mst2(
  dom,
  y,
  x,
  index_d,
  vardir,
  graph0,
  betam0,
  lamvec,
  rho = 1,
  gam = 3,
  lam0 = 0.001,
  maxiter = 500,
  tol = 1e-04
)
}
\description{
based on given data, give the area level model estimator new algorithm
}
