% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_clustSAR_area_mst.R
\name{Est_clustSAR_area_mst}
\alias{Est_clustSAR_area_mst}
\title{based on given data, give the area level model estimator with SAR}
\usage{
Est_clustSAR_area_mst(
  dom,
  y,
  x,
  index_d,
  vardir,
  graph0,
  betam0,
  tau0 = NULL,
  rho0 = NULL,
  adjW,
  lamvec,
  nu = 1,
  gam = 3,
  lam0 = 0.001,
  maxiter = 500,
  tol = 1e-04
)
}
\description{
based on given data, give the area level model estimator with SAR
}
