% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_clust_area.R
\name{fit_clust_area}
\alias{fit_clust_area}
\title{fit the model based on a give cluster structure and given vardir ####}
\usage{
fit_clust_area(dom, y, x, index_d, vardir, cluster)
}
\description{
fit the model based on a give cluster structure and given vardir ####
}
